# evechem_api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0"
  title: "Evechem API"
schemes:
- "https"
paths:
  /operation/:
    get:
      tags:
      - "Operation"
      description: "Gets general information about the operation accessed by the given\
        \ master key.\n"
      operationId: "operation_get"
      parameters:
      - name: "api_key"
        in: "query"
        description: "Operation Master Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/operation"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.operation_controller"
    post:
      tags:
      - "Operation"
      description: "Creates a new operation and returns operation details (including\
        \ the `master key`). Optional parameters include the private name and public\
        \ name of the operation.  These can be changed later.\n"
      operationId: "operation_post"
      parameters:
      - in: "body"
        name: "operation_name"
        description: "Operation Name and Public Name"
        required: false
        schema:
          $ref: "#/definitions/operation_name"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/operation"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.operation_controller"
    delete:
      tags:
      - "Operation"
      description: "Deletes the operation accessed by the given master key.  Note\
        \ that this invalidates the key and deletes _all_ information under the operation\
        \ including api keys, towers, and thier configurations.\n"
      operationId: "operation_delete"
      parameters:
      - name: "api_key"
        in: "query"
        description: "Operation Master Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.operation_controller"
    patch:
      tags:
      - "Operation"
      description: "Update either private or public name of the operation accessed\
        \ with the provided master key.\n"
      operationId: "operation_patch"
      parameters:
      - name: "api_key"
        in: "query"
        description: "Operation Master Access Key"
        required: true
        type: "string"
      - in: "body"
        name: "operation_name"
        description: "Operation Name and Public Name"
        required: true
        schema:
          $ref: "#/definitions/operation_name"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/operation"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.operation_controller"
  /operation/keys/:
    get:
      tags:
      - "Operation"
      description: "Returns `array` of operation access sub-keys that exist for this\
        \ operation.\n"
      operationId: "operation_keys_get"
      parameters:
      - name: "api_key"
        in: "query"
        description: "Operation Master Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            title: "access_keys"
            items:
              type: "string"
              title: "access_key"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.operation_controller"
    post:
      tags:
      - "Operation"
      description: "Creates a new operation access sub-key with the access scopes\
        \ provided, and then returns the key in the response.\n"
      operationId: "operation_keys_post"
      parameters:
      - name: "api_key"
        in: "query"
        description: "Operation Master Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "string"
            title: "access_key"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.operation_controller"
  /operation/keys/{sub_key}/:
    get:
      tags:
      - "Operation"
      description: "Gets the permission settings of the given access key.\n"
      operationId: "operation_keys_sub_key_get"
      parameters:
      - name: "sub_key"
        in: "path"
        description: "Operation Access Sub-key"
        required: true
        type: "string"
      - name: "api_key"
        in: "query"
        description: "Operation Master Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/key_type"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Key Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.operation_controller"
    delete:
      tags:
      - "Operation"
      description: "Delete an access sub-key.  The specified key will no longer be\
        \ valid for accessing any portion of the given operation.\n"
      operationId: "operation_keys_sub_key_delete"
      parameters:
      - name: "sub_key"
        in: "path"
        description: "Operation Access Sub-key"
        required: true
        type: "string"
      - name: "api_key"
        in: "query"
        description: "Operation Master Access Key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Key Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.operation_controller"
    patch:
      tags:
      - "Operation"
      description: "Update key_type for the given access sub-key.\n"
      operationId: "operation_keys_sub_key_patch"
      parameters:
      - name: "sub_key"
        in: "path"
        description: "Operation Access Sub-key"
        required: true
        type: "string"
      - name: "api_key"
        in: "query"
        description: "Operation Master Access Key"
        required: true
        type: "string"
      - in: "body"
        name: "key_type"
        required: true
        schema:
          $ref: "#/definitions/key_type"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/key_type"
        401:
          description: "Authentication Required"
          schema:
            $ref: "#/definitions/error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/error"
        404:
          description: "Key Not Found"
          schema:
            $ref: "#/definitions/error"
      #security:
      #- api_key: []
      x-swagger-router-controller: "evechem_api.controllers.operation_controller"
...
definitions:
...
  error:
    type: "object"
    properties:
      message:
        type: "string"
    title: "error"
...
  key_type:
    type: "string"
  operation:
    type: "object"
    properties:
      master_key:
        type: "string"
        description: "Master Access Key for this Operation"
      name:
        type: "string"
        description: "Private Name (not viewable by `customer` type keys)"
      public_name:
        type: "string"
        description: "Name viewable by `customer` type keys"
      tower_count:
        type: "integer"
        description: "Total operation tower count"
      sub_key_count:
        type: "integer"
        description: "Total sub-keys active for this operation"
  operation_name:
    type: "object"
    properties:
      name:
        type: "string"
      public_name:
        type: "string"